# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "წარმატებული რეგისტრაცია",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          },
          "409": {
            "description": "ელფოსტა უკვე დაკავებულია"
          }
        },
        "summary": "მომხმარებლის რეგისტრაცია",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "წარმატებული ავტორიზაცია",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ვალიდაციის შეცდომა"
          },
          "401": {
            "description": "არასწორი ელფოსტა ან პაროლი"
          }
        },
        "summary": "მომხმარებლის ავტორიზაცია",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "წარმატებული წვდომა"
          },
          "401": {
            "description": "არ ხართ ავტორიზებული"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "მიმდინარე მომხმარებლის პროფილი",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/dashboard": {
      "get": {
        "operationId": "AuthController_getAdminDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "წარმატებული წვდომა"
          },
          "401": {
            "description": "არ ხართ ავტორიზებული"
          },
          "403": {
            "description": "არ გაქვს ადმინისტრატორის უფლებები"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "მხოლოდ ადმინისტრატორის პანელი",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "პაროლი წარმატებით შეიცვალა",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ძველი პაროლი არასწორია ან ახალი პაროლი ძველს ემთხვევა"
          },
          "401": {
            "description": "არ ხართ ავტორიზებული"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "პაროლის შეცვლა (ავტორიზებული მომხმარებლისთვის)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/google": {
      "post": {
        "operationId": "AuthController_googleLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "წარმატებული ავტორიზაცია",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "Google-ით ავტორიზაცია/რეგისტრაცია",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ინსტრუქცია გაიგზავნა"
          },
          "400": {
            "description": "არასწორი ელფოსტა"
          }
        },
        "summary": "პაროლის აღდგენის მოთხოვნა (email-ის გაგზავნა)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "პაროლი წარმატებით შეიცვალა"
          },
          "400": {
            "description": "არასწორი ან ამოწურული token"
          }
        },
        "summary": "პაროლის აღდგენა token-ით",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Online Shop API",
    "description": "REST API for the online shop backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "gender": {
            "enum": [
              "male",
              "female"
            ],
            "type": "string",
            "description": "მომხმარებლის სქესი"
          },
          "age": {
            "type": "number",
            "minimum": 0,
            "description": "მომხმარებლის ასაკი"
          },
          "personalNumber": {
            "type": "string",
            "pattern": "^\\d{11}$",
            "description": "პირადი ნომერი — ზუსტად 11 ციფრი",
            "example": "01234567890"
          },
          "otpRequestId": {
            "type": "string",
            "description": "POST /otp/send-ის პასუხიდან მიღებული requestId — თუ SMS ვერიფიკაცია ჩართულია, სავალდებულოა otpCode-თან ერთად"
          },
          "otpCode": {
            "type": "string",
            "description": "მომხმარებლის მობილურზე მიღებული OTP კოდი"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "phoneNumber"
        ]
      },
      "TokenUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "firstName": {
            "type": "string",
            "example": "გიორგი"
          },
          "lastName": {
            "type": "string",
            "example": "ბერიძე"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string",
            "example": "user"
          },
          "gender": {
            "enum": [
              "male",
              "female"
            ],
            "type": "string"
          },
          "age": {
            "type": "number",
            "example": 25
          },
          "personalNumber": {
            "type": "string",
            "example": "*********90"
          },
          "phoneNumber": {
            "type": "string",
            "example": "*********3456"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": false
          },
          "isPhoneVerified": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "phoneNumber",
          "isEmailVerified",
          "isPhoneVerified"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/TokenUserDto"
          }
        },
        "required": [
          "access_token",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/TokenUserDto"
          }
        },
        "required": [
          "access_token",
          "user"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "ძველი პაროლი",
            "example": "oldPassword123"
          },
          "newPassword": {
            "type": "string",
            "description": "ახალი პაროლი (მინიმუმ 8 სიმბოლო, დიდი და პატარა ასო და ციფრი)",
            "example": "newPassword456"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "ChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "პაროლი წარმატებით შეიცვალა"
          },
          "access_token": {
            "type": "string",
            "description": "ახალი access token — ძველი ამ მოთხოვნის შემდეგ ბათილია",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "statusCode",
          "message",
          "access_token"
        ]
      },
      "GoogleLoginDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google Identity Services-ის მიერ გაცემული ID Token (JWT)"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "dato@gmail.com",
            "description": "მომხმარებლის ელფოსტა"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token, რომელიც მიღებულია email-იდან"
          },
          "newPassword": {
            "type": "string",
            "description": "ახალი პაროლი",
            "example": "newPassword123"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
...
    }
  }
}